+++ test/Fixture/SchemaNormalizer/NormalizeNormalizes/WithCustomJsonPointerSpecification/Data/IsObject/Schema/HasType/IsScalar/WithPropertyDefinitions/specification-for-pointer-to-data-that-should-not-be-sorted.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2018-2022 Andreas Möller
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE.md file that was distributed with this source code.
+ *
+ * @see https://github.com/ergebnis/json-normalizer
+ */
+
+use Ergebnis\Json\Pointer;
+
+return Pointer\Specification::equals(Pointer\JsonPointer::fromJsonString('/bar/baz'));