+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For a full diff see [`3.0.0...main`][3.0.0...main].
 
 - Dropped support for PHP 7.4 ([#757]), by [@localheinz]
 - Required `ergebnis/json-schema-validator:^4.0.0` ([#771]), by [@localheinz]
+- Allowed configuring the `Normalizer\SchemaNormalizer` to exclude properties from being sorted ([#774]), by [@localheinz]
 
 ### Removed
 
@@ -520,6 +521,7 @@ For a full diff see [`5d8b3e2...0.1.0`][5d8b3e2...0.1.0].
 [#723]: https://github.com/ergebnis/json-normalizer/pull/723
 [#757]: https://github.com/ergebnis/json-normalizer/pull/757
 [#772]: https://github.com/ergebnis/json-normalizer/pull/772
+[#774]: https://github.com/ergebnis/json-normalizer/pull/774
 
 [@BackEndTea]: https://github.com/BackEndTea
 [@dependabot]: https://github.com/dependabot